1 error in unsigned and 5 in ptr

+++ src/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: atweek <atweek@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/08 15:55:23 by atweek            #+#    #+#             */
-/*   Updated: 2021/01/20 20:47:02 by atweek           ###   ########.fr       */
+/*   Updated: 2021/01/21 15:34:42 by atweek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ int					ft_intlen(int num);
 int x_processing(t_pars *st_pars, va_list argptr);
 int					num_zero(t_pars *st_pars);
 int hex_convert(unsigned int hex,int len,t_pars *st_pars);
-int hex_intlen(int num);
+int hex_intlen(unsigned long long int num);
 int p_processing(t_pars *st_pars, va_list argptr);
